refactor how tests are run and handle the vault server subprocess
