+++ ui/package.json
@@ -10,20 +10,20 @@
   },
   "scripts": {
     "build": "ember build -prod",
-    "build-dev": "ember build",
+    "build:dev": "ember build",
     "lint:hbs": "ember-template-lint app/**/* lib/**/*",
     "lint:js": "eslint .",
     "fmt": "yarn run fmt-js && yarn run fmt-styles",
-    "fmt-js": "prettier-eslint --single-quote --no-use-tabs --trailing-comma es5 --print-width=110 --write '{app,tests,config,lib}/**/*.js'",
-    "fmt-styles": "prettier --write app/styles/**/*.*",
+    "fmt:js": "prettier-eslint --single-quote --no-use-tabs --trailing-comma es5 --print-width=110 --write '{app,tests,config,lib}/**/*.js'",
+    "fmt:styles": "prettier --write app/styles/**/*.*",
     "start": "export VAULT_ADDR=http://localhost:8200; ember server --proxy=$VAULT_ADDR",
     "start2": "ember server --proxy=http://localhost:8202 --port=4202",
-    "test": "node scripts/start-vault.js & yarn run lint:js & ember test",
-    "test:browserstack": "export CI=true; node scripts/start-vault.js & node scripts/run-browserstack-tests.js",
-    "test-oss": "yarn run test -f='!enterprise'",
-    "test-quick": "node scripts/start-vault.js & ember test",
-    "test-quick-oss": "yarn run test-quick -f='!enterprise'",
-    "build-storybook": "build-storybook -s ../pkg/web_ui",
+    "test": "yarn lint:js && node scripts/start-vault.js",
+    "test:oss": "yarn run test -f='!enterprise'",
+    "test:browserstack": "export CI=true; node scripts/start-vault.js --browserstack",
+    "test:quick": "node scripts/start-vault.js",
+    "test:quick-oss": "yarn test-quick -f='!enterprise'",
+    "build:storybook": "build-storybook -s ../pkg/web_ui",
     "storybook": "start-storybook -p 6006 -s ../pkg/web_ui",
     "gen-story-md": "node scripts/gen-story-md.js"
   },
@@ -114,8 +114,8 @@
     "escape-string-regexp": "^2.0.0",
     "eslint": "^6.1.0",
     "eslint-config-prettier": "^6.0.0",
-    "eslint-plugin-ember": "^6.3.0",
-    "eslint-plugin-prettier": "^3.0.0",
+    "eslint-plugin-ember": "^6.7.0",
+    "eslint-plugin-prettier": "^3.1.0",
     "flat": "^4.1.0",
     "ivy-codemirror": "IvyApp/ivy-codemirror#fb09333c5144da47e14a9e6260f80577d5408374",
     "jsonlint": "1.6.0",
