+++ ui/scripts/start-vault.js
@@ -1,80 +1,98 @@
 #!/usr/bin/env node
-/* eslint-disable */
+/* eslint-env node */
+/* eslint-disable no-console */
 
-if (process.argv[2]) {
-  process.kill(process.argv[2], 'SIGINT');
-  process.exit(0);
-}
-
-process.env.TERM = 'dumb';
 var fs = require('fs');
 var path = require('path');
 var readline = require('readline');
-var spawn = require('child_process').spawn;
-var vault = spawn('vault', [
-  'server',
-  '-dev',
-  '-dev-ha',
-  '-dev-transactional',
-  '-dev-root-token-id=root',
-  '-dev-listen-address=127.0.0.1:9200',
-]);
+var execa = require('execa');
+var chalk = require('chalk');
+
+function run(command, args = []) {
+  console.log(chalk.dim('$ ' + command + ' ' + args.join(' ')));
+  // cleanup means that execa will handle stopping the vault subprocess
+  // inherit all of the stdin/out/err so that testem still works as if you were running it directly
+  return execa(command, args, { cleanup: true, stdin: 'inherit', stdout: 'inherit', stderr: 'inherit' });
+}
 
 var output = '';
-var unseal, root;
+var unseal, root, written;
 
-readline
-  .createInterface({
-    input: vault.stdout,
-    terminal: false,
-  })
-  .on('line', function(line) {
-    output = output + line;
-    console.log(line);
-    var unsealMatch = output.match(/Unseal Key: (.+)$/m);
-    if (unsealMatch && !unseal) {
-      unseal = unsealMatch[1];
-    }
-    var rootMatch = output.match(/Root Token: (.+)$/m);
-    if (rootMatch && !root) {
-      root = rootMatch[1];
-    }
-    if (root && unseal) {
-      fs.writeFile(
-        path.join(process.cwd(), 'tests/helpers/vault-keys.js'),
-        `export default ${JSON.stringify({ unseal, root }, null, 2)}`,
-        err => {
-          if (err) throw err;
-        }
-      );
-
-      console.log('VAULT SERVER READY');
-    }
+async function processLines(input, eachLine = () => {}) {
+  const rl = readline.createInterface({
+    input,
+    terminal: true,
   });
+  for await (const line of rl) {
+    eachLine(line);
+  }
+}
 
-vault.stderr.on('data', function(data) {
-  console.log(data.toString());
-});
+(async function() {
+  try {
+    run(
+      'vault',
+      [
+        'server',
+        '-dev',
+        '-dev-ha',
+        '-dev-transactional',
+        '-dev-root-token-id=root',
+        '-dev-listen-address=127.0.0.1:9200',
+      ],
+      false
+    );
 
-vault.on('close', function(code) {
-  console.log(`child process exited with code ${code}`);
-  process.exit();
-});
-vault.on('error', function(error) {
-  console.log(`child process errored: ${error}`);
-  process.exit();
-});
+    processLines(process.stdout, function(line) {
+      if (written) {
+        output = null;
+        return;
+      }
+      output = output + line;
+      var unsealMatch = output.match(/Unseal Key: (.+)$/m);
+      if (unsealMatch && !unseal) {
+        unseal = unsealMatch[1];
+      }
+      var rootMatch = output.match(/Root Token: (.+)$/m);
+      if (rootMatch && !root) {
+        root = rootMatch[1];
+      }
+      if (root && unseal && written !== true) {
+        fs.writeFile(
+          path.join(process.cwd(), 'tests/helpers/vault-keys.js'),
+          `export default ${JSON.stringify({ unseal, root }, null, 2)}`,
+          err => {
+            if (err) throw err;
+          }
+        );
+        written = true;
+        console.log('VAULT SERVER READY');
+      }
+    });
+    try {
+      if (process.argv[2] === '--browserstack') {
+        await run('ember', ['browserstack:connect']);
+        try {
+          await run('ember', ['test', '-f=secrets/secret/create', '-c', 'testem.browserstack.js']);
 
-var pidFile = 'vault-ui-integration-server.pid';
-process.on('SIGINT', function() {
-  vault.kill('SIGINT');
-  process.exit();
-});
-process.on('exit', function() {
-  vault.kill('SIGINT');
-});
-
-fs.writeFile(pidFile, process.pid, err => {
-  if (err) throw err;
-  console.log('The file has been saved!');
-});
+          console.log('success');
+          process.exit(0);
+        } finally {
+          if (process.env.CI === 'true') {
+            await run('ember', ['browserstack:results']);
+          }
+          await run('ember', ['browserstack:disconnect']);
+        }
+      } else {
+        await run('ember', ['test', ...process.argv.slice(2)]);
+      }
+    } catch (error) {
+      console.log(error);
+    } finally {
+      process.exit(0);
+    }
+  } catch (error) {
+    console.log(error);
+    process.exit(0);
+  }
+})();
