+++ ui/testem.js
@@ -18,9 +18,6 @@ const config = {
       ].filter(Boolean),
     },
   },
-  on_exit:
-    '[ -e ../../vault-ui-integration-server.pid ] && node ../../scripts/start-vault.js `cat ../../vault-ui-integration-server.pid`; [ -e ../../vault-ui-integration-server.pid ] && rm ../../vault-ui-integration-server.pid',
-
   proxies: {
     '/v1': {
       target: 'http://localhost:9200',
